Optimize, better parallelize the Parse-Annotate-Store (PAS) pipeline
The gain from the PAS parallelization is small (low single digit %)
but this split with a channel for communication is a natural
progressing for such a tool.

The bottleneck at this point is the data store
(60-67% depending on the input size and I/O capacity
available).

A number of SQLite tweaks and some trial-and-error
provide a 13-16% gain:

 * synchronous=OFF to greatly reduce the number of fsync(2) calls
 * journal_mode=WAL
 * an increased cache size
 * temp_store=MEMORY